requirements.txt
todoist-api-python
python-dotenv
pre-commit
